made displaymarkers a global function in map.js so it can be reused/called in other code parts.
